v1/labs/requirements.txt
# Microsoft AutoGen Course - Dependencies
pyautogen>=0.2.0
openai>=1.0.0
python-dotenv>=1.0.0
